Create war_of_species.py
An empty cell (".") becomes non-empty if it is surrounded by at least two
individuals of the same species. In particular, it becomes an individual
of the most frequent species in its neighbourhood. In case of a draw
between the species, the cell remains empty.
• A non-empty cell becomes empty if it is surrounded by more than six
non-empty cells, regardless of their species.
• A non-empty cell becomes empty if it is surrounded by less than three
members of its species.
• A non-empty cell becomes empty if it is surrounded by more members
of the opposite species than members of its species.
• In any other circumstances, a cell does not change its value.
You can assume that the format of the input list of strings is correct.
Example: The initial environment
["XX......", "XX....O.", ".....OOO"]
becomes
["XXX.....", "XXX..OOO", "XX....O."]